1216 끝

+++ src/main/java/com/app/home/messenger/NoteVO.java
@@ -9,8 +9,8 @@ import lombok.Data;
 public class NoteVO {
 	
 	private Long noteNum;
-	private Long sendId;
-	private Long receiveId;
+	private int sendId;
+	private int receiveId;
 	private String contents;
 	private Timestamp regDate;
 	private int readCheck;
